pyproject.toml
[tool.poetry]
name = "archiver"
version = "0.1.0"
description = "vore"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.24.1"
py-cord = "^2.4.1"
gallery-dl = "^1.25.8"
yt-dlp = "^2023.7.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
